Move `TodoList` type to own file
