+++ src/components/todos/todo-list-dialog.tsx
@@ -10,7 +10,7 @@ import {
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
-import { TodoList } from '@/types';
+import { TodoList } from '@/types/todo-list';
 import { LoaderIcon, X } from 'lucide-react';
 import { forwardRef, useImperativeHandle, useRef, useState } from 'react';
 import { toast } from 'sonner';
